backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
Pillow
pydantic
aiosqlite
